combined all different methods into 1 file
Easier to maintain, less code dupe
